#214 build(fe): 로직 분리
